[TASK] Add support for CollectionStringCastTransformer
Currently the implementation doesn't support property with type array. I
have a small path to add a transformer class to cast collection objects
to string. A small modification need to be done in the
objectIndexActionRequired, because array_diff_assoc is used to compare
the two array. This function has no support for multi dimensional array

Change-Id: I7c23592a1107814fe3fe3d39106aecde1c3003af
Resolves: #47499