[BUGFIX] Request drops username / password in some cases
This change makes the RequestService more resilient for some cases when
the HTTP Request sent to Elasticsearch did not contain the username
and password anymore, even though it has been configured in the
settings.

While the root cause is the Http\Request::create() method in Flow,
this change at least provides a workaround which allows for connecting
with Elasticsearch instances protected with HTTP Basic Auth.

Also circumvents a rare case where the content of a PUT request contains
"null" as a string.