TASK: Avoid double indexing of entities
We index via AOP and via Doctrine lifecycle events. This may lead to double
indexing.

There was code in place to avoid double indexing, but that was removed as some
point, see https://github.com/Flowpack/Flowpack.ElasticSearch/pull/36.

This change brings back a check for objects that were already handled, based
on the persistence identifier that is used anyway.